# swagger.yaml
openapi: 3.0.3
info:
  title: Mosaic API
  description: "Mosaic will be a social media platform designed for individuals to\
    \ engage in discussions and share content in a community-driven environment. Similar\
    \ to Reddit, Mosaic will allow users to create and join communities that are focused\
    \ on specific topics of interest. As the name suggests, our aim is to bring together\
    \ a diverse community made up of different pieces that come together to create\
    \ a beautiful and meaningful whole."
  termsOfService: http://link_to_mosaic/terms/
  contact:
    email: apiteam@mosaic.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: "1.0"
...
servers:
- url: https://link_to_mosaic/api/v3
...
paths:
...
  /c/create:
    post:
      tags:
      - communities
      summary: Create community
      description: This can only be done by the logged in user.
      operationId: create_community
      requestBody:
        description: Created community object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Communities'
        required: true
      responses:
        default:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Communities'
      security:
      - user_auth:
        - login:user
      x-openapi-router-controller: swagger_server.controllers.communities_controller
  /c/{community_name}:
    get:
      tags:
      - communities
      summary: Get community by name
      description: ""
      operationId: get_community_by_name
      parameters:
      - name: community_name
        in: path
        description: 'The name of the community. Use askMosaic for testing. '
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Communities'
        "400":
          description: Invalid community name supplied
        "404":
          description: Community not found
      x-openapi-router-controller: swagger_server.controllers.communities_controller
    put:
      tags:
      - communities
      summary: Update community
      description: This can only be done by the logged in user who is either an admin
        or moderator of the community.
      operationId: update_community
      parameters:
      - name: community_name
        in: path
        description: 'The name of the community. Use askMosaic for testing. '
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        description: Update an existent community
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        default:
          description: successful operation
      security:
      - user_auth:
        - login:user
      x-openapi-router-controller: swagger_server.controllers.communities_controller
    delete:
      tags:
      - communities
      summary: Delete community
      description: This can only be done by the logged in user who is the admin of
        the community.
      operationId: delete_community
      parameters:
      - name: community_name
        in: path
        description: 'The name of the community. Use askMosaic for testing. '
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "400":
          description: Invalid community name supplied
        "404":
          description: Community not found
      security:
      - user_auth:
        - login:user
      x-openapi-router-controller: swagger_server.controllers.communities_controller
...
  /p/create:
    post:
      tags:
      - posts
      summary: Create Post
      description: This can only be done by the logged in user.
      operationId: create_post
      requestBody:
        description: Created posted object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Posts'
        required: true
      responses:
        default:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Posts'
      security:
      - user_auth:
        - login:user
      x-openapi-router-controller: swagger_server.controllers.posts_controller
  /p/{postID}:
    get:
      tags:
      - posts
      summary: Get post by id
      description: ""
      operationId: get_post_by_id
      parameters:
      - name: postID
        in: path
        description: The ID of the post
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Posts'
        "400":
          description: Invalid post id supplied
        "404":
          description: Post not found
      x-openapi-router-controller: swagger_server.controllers.posts_controller
    put:
      tags:
      - posts
      summary: Update post
      description: This can only be done by the logged in user who is the creator
        of the post.
      operationId: update_post
      parameters:
      - name: postID
        in: path
        description: The ID of the post
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      requestBody:
        description: Update an existent post
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Posts'
      responses:
        default:
          description: successful operation
      security:
      - user_auth:
        - login:user
      x-openapi-router-controller: swagger_server.controllers.posts_controller
    delete:
      tags:
      - posts
      summary: Delete post
      description: This can only be done by the logged in user who created the post
        or the one who is either a moderator or admin of the community.
      operationId: delete_post
      parameters:
      - name: postID
        in: path
        description: The ID of the post
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      responses:
        "400":
          description: Invalid post id supplied
        "404":
          description: Post not found
      security:
      - user_auth:
        - login:user
      x-openapi-router-controller: swagger_server.controllers.posts_controller
...
  /u/create:
    post:
      tags:
      - user
      summary: Create user
      description: This can only be done by the logged in user.
      operationId: create_user
      requestBody:
        description: Created user object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        default:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
      x-openapi-router-controller: swagger_server.controllers.user_controller
  /u/login:
    post:
      tags:
      - user
      summary: Logs user into the system
      description: ""
      operationId: login_user
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/u_login_body'
        required: true
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: string
                x-content-type: application/json
        "400":
          description: Invalid username/password supplied
      x-openapi-router-controller: swagger_server.controllers.user_controller
...
  /u/{username}:
    get:
      tags:
      - user
      summary: Get user by user name
      description: ""
      operationId: get_user_by_name
      parameters:
      - name: username
        in: path
        description: The name of the user. Use reaper_113 for testing.
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        "400":
          description: Invalid username supplied
        "404":
          description: User not found
      x-openapi-router-controller: swagger_server.controllers.user_controller
    put:
      tags:
      - user
      summary: Update user
      description: This can only be done by the logged in user.
      operationId: update_user
      parameters:
      - name: username
        in: path
        description: The name of the user. Use reaper_113 for testing.
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        description: Update an existent user in the store
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        default:
          description: successful operation
      security:
      - user_auth:
        - login:user
      x-openapi-router-controller: swagger_server.controllers.user_controller
    delete:
      tags:
      - user
      summary: Delete user
      description: This can only be done by the logged in user.
      operationId: delete_user
      parameters:
      - name: username
        in: path
        description: The name of the user. Use reaper_113 for testing.
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "400":
          description: Invalid username supplied
        "404":
          description: User not found
      security:
      - user_auth:
        - login:user
      x-openapi-router-controller: swagger_server.controllers.user_controller
...
  /comments:
    post:
      tags:
      - comments
      summary: Add comment
      description: This can only be done by the logged in user.
      operationId: create_comment
      requestBody:
        description: Created commented object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Comments'
        required: true
      responses:
        default:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comments'
      security:
      - user_auth:
        - login:user
      x-openapi-router-controller: swagger_server.controllers.comments_controller
...
components:
  schemas:
...
    u_login_body:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
...
  securitySchemes:
    user_auth:
      type: oauth2
      flows:
        password:
          tokenUrl: https://link_to_mosaic.com/oauth/authorize
          scopes:
            login:user: user logged in
      x-tokenInfoFunc: swagger_server.controllers.authorization_controller.check_user_auth
      x-scopeValidateFunc: swagger_server.controllers.authorization_controller.validate_scope_user_auth
    api_key:
      type: apiKey
      name: api_key
      in: header
      x-apikeyInfoFunc: swagger_server.controllers.authorization_controller.check_api_key